src/shared/lib/url/addQueryParams/addQueryParams.test.ts
import { getQueryParams } from './addQueryParams';

describe('shared/url/addQueryParams', () => {
    test('test один параметр', () => {
        const params = getQueryParams({
            test: 'value',
        });
        expect(params).toBe('?test=value');
    });
    test('test несколько параметров', () => {
        const params = getQueryParams({
            test: 'value',
            second: '2',
        });
        expect(params).toBe('?test=value&second=2');
    });
    test('test параметр === undefined', () => {
        const params = getQueryParams({
            test: 'value',
            second: undefined,
        });
        expect(params).toBe('?test=value');
    });
});
